tersoff1988 single precision

+++ src/force/tersoff1988.cuh
@@ -19,8 +19,8 @@
 #include <stdio.h>
 
 struct Tersoff1988_Data {
-  GPU_Vector<double> b;   // bond orders
-  GPU_Vector<double> bp;  // derivative of bond orders
+  GPU_Vector<float> b;    // bond orders
+  GPU_Vector<float> bp;   // derivative of bond orders
   GPU_Vector<float> f12x; // partial forces
   GPU_Vector<float> f12y;
   GPU_Vector<float> f12z;
@@ -50,6 +50,6 @@ public:
 
 protected:
   int num_types;
-  GPU_Vector<double> ters;
+  GPU_Vector<float> ters;
   Tersoff1988_Data tersoff_data;
 };
